Queue requests for titles by author
